Move more stuff to handlers
